Swipe to Complete (#39)
* Initial swipe to complete

Swipes to complete task. So far, it just removes it from the list. Also improved transparency visual effect of cards (shows waves beneath).

* Cool Colors

Subjects are now sorted by colors as defined by a dictionary in UserData. Missing method to create a new subject when adding a To-Do item (to be implemented).

* Assigned RandomColor to Swift 4.2